Practice code added into the repo

+++ mapping/src/main/java/com/sesc/mapping/repositories/AddressRepository.java
@@ -1,7 +1,7 @@
 package com.sesc.mapping.repositories;
 
-import com.sesc.mapping.entities.Address;
+import com.sesc.mapping.models.Address;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AddressRepository extends JpaRepository<Address,Long> {
+public interface AddressRepository extends JpaRepository<Address,Integer> {
 }
