+++ mapping/src/main/java/com/sesc/mapping/repositories/StudentRepository.java
@@ -1,7 +1,7 @@
 package com.sesc.mapping.repositories;
 
-import com.sesc.mapping.entities.Student;
+import com.sesc.mapping.models.Student;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface StudentRepository extends JpaRepository<Student,Long> {
+public interface StudentRepository extends JpaRepository<Student,Integer> {
 }
